+++ lib/Imagine/Image/Palette/RGB.php
@@ -117,13 +117,16 @@ class RGB implements PaletteInterface
             throw new RuntimeException('RGB palette can only blend RGB colors');
         }
 
+        $amount = (float) $amount;
+        $amountComplement = 1 - $amount;
+
         return $this->color(
             array(
-                (int) min(255, min($color1->getRed(), $color2->getRed()) + round(abs($color2->getRed() - $color1->getRed()) * $amount)),
-                (int) min(255, min($color1->getGreen(), $color2->getGreen()) + round(abs($color2->getGreen() - $color1->getGreen()) * $amount)),
-                (int) min(255, min($color1->getBlue(), $color2->getBlue()) + round(abs($color2->getBlue() - $color1->getBlue()) * $amount)),
+                min(max((int) round($color2->getRed() * $amount + $color1->getRed() * $amountComplement), 0), 255),
+                min(max((int) round($color2->getGreen() * $amount + $color1->getGreen() * $amountComplement), 0), 255),
+                min(max((int) round($color2->getBlue() * $amount + $color1->getBlue() * $amountComplement), 0), 255),
             ),
-            (int) min(100, min($color1->getAlpha(), $color2->getAlpha()) + round(abs($color2->getAlpha() - $color1->getAlpha()) * $amount))
+            min(max((int) round($color2->getAlpha() * $amount + $color1->getAlpha() * $amountComplement), 0), 255)
         );
     }
 }