+++ CHANGELOG.md
@@ -41,6 +41,7 @@
     **NOTE** the `$mode` argument has been renamed to `$settings`, and it's now an integer (but old string values are accepted for backward compatibility)
   * New filter: `brightness` (@lenybernard, @mlocati)
   * New filter: `colvolve` available for all graphics libraries except gmagick with version prior to 2.0.1RC2 (@armatronic, @mlocati)
+  * Fix bug in Imagine\Image\Palette\RGB::blend() (@dmolineus, @mlocati)
 
 ### 0.7.1 (2017-05-16)
   * Remove Symfony PHPUnit bridge as dependency (@craue)